Support Loading Metrics Configs from Labels
Previously, the custom metrics configuration was loaded
from a file in the container, by specifiying a file path
as the value for a label of the form "io.cadvisor.metrics.xyz".

Now, you can also just specify the configuration JSON directly
in label, by using a label with the name
"io.cadvisor.metric-config.xyz".

Part of #1016