Extract FS caching into PartitionCache
Currently, the FsInfo module caches all mounts at startup, and never
refreshes its cache, which means volumes that are mounted later on will
be missed.

This updates FsInfo to use a separate PartitionCache that is refreshed
when the PartitionCache is queried for information it doesn't have but
is likely to be the result of a stale cache.