Initial setup for building aarch64
This patch modifies the build scripts to allow other
architectures such as aarch64 to be built.

Particularly, for the aarch64 case, the following changes are included:

- gccgo-aarch64-linux-gnu: is required for building
aarch64 binaries , added as the default CC for arm64.

- Added a Dockerfile under deploy/Dockerfile.arm64 that
uses multiarch/alpine:aarch64-edge as the base image.

- Modified the Makefiles and build/*.sh scripts to allow
the ARCH variable to be passed.

Signed-off-by: Jorge Niedbalski <jnr@metaklass.org>