Add flag to white list container labels for prometheus metrics
This adds a flag `white_listed_container_labels` which allows specifying
a comma separated list of container labels that are allowed to be
attached to prometheus metrics. This helps limit the label space of
each data point and limits the memory used by cadvisor compared to just
add all labels from all containers which is the default.

This is somewhat similar to the `store_container_labels` flag but a bit
more flexible because it allows white listing some common labels e.g.
`io.kubernetes.pod.name`, `io.kubernetes.pod.namespace` etc.

Without this white listing in place we saw the memory footprint of
cadvisor explode in our production environment because users add a lot
of data in pod labels, data which is not useful for metrics but may be
useful for the individual applications.

Signed-off-by: Mikkel Oscar Lyderik Larsen <mikkel.larsen@zalando.de>