Issue fix: get wrong machine info in Arm64 guest
The latest kubernetes deployment on Arm64 VM-s always fails.
Because k8s always get num_cores=0 from cAdvisor on Arm64 VM-s.

The reason is that, there is no cache info on Arm64 VM-s.
And the good news is that, we can get cache info on Arm64 hosts.

When this patch was merged, I will deliver a patch to update the version
of cAdvisor in kubernetes as soon as possible.

Signed-off-by: bblu <bin.lu@arm.com>