add option to disable socket count collection
Socket count costs a lot of CPU cycles to collect as we need to perform readlink
to determine if fd is a socket. In many applications/use cases
such as load balancers etc, just fd count is enough for many monitoring purposes,
so we add a option to disable socket count stats collection here with
`process_socket_count` metric.

Signed-off-by: Daniel Dao <dqminh89@gmail.com>