Clarify MetricSet thread-safety status
Both Add() and Set() methods modify the existing mapping.

As Add() method is used only by Set() and Difference(), and
in latter it's done thread-safely, make add private method
and document that Set() is not thread-safe.

Set() is used only by flag package through Value interface,
which is documented to be called only once, i.e. it's safe.