container: rm cgroupFilesystems from rawContainerWatcher
cgroupPaths was basically a copy of cgroupSubsystems.MountPoints.
This is because it was produced from the elements of the latter
by calling common.MakeCgroupPaths() with the second argument of "/".
Note that path.Join(something, "/") will return something.

Since we have two copies of the same info, let's remove one and reuse
the other.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>