container/libcontainer: simplify for cgroup v2
In case of cgroup v2, there are no separate per-subsystem mount points,
and so it does not make sense to have a subsystem to mountpoint map.

Yet for the unification of code between v1 and v2, we still use this
map, although a single element is enough.

For cgroup v2, the key is "", the value is the mount point (or the path
to a specific cgroup).

This commit
 - simplifes GetCgroupSubsystems for cgroup v2 (trivial) case;
 - modifies GetCgroupPath() methods to use "" as the key for cgroup v2;
 - fixes NewCgroupManager() to use "" as the key for cgroup v2 case.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>