Makefile: install golangci-lint if not found
Check if golangci-lint is available locally -- if not, install it
to GOPATH/bin. This helps with cases when golangci-lint is not
available on the host (such as in prow).

As we now have golangci-lint installation in Makefile, remove the
code installing it from .github/workflows/test.yml for better
maintainability.

While at it, add lint to .PHONY targets.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>