manager: fix determining selfContainer name
In the cgroup v1 case, the `selfContainer` variable is shadowed in the block
and does not update the variable outside the block. This can be observed by
commenting out the "cAdvisor running in container:" log statement, which
results in a compiler error:

  ../manager/manager.go:161:3: selfContainer declared but not used

This resulted in the selfContainer variable always having the value '/' later
in the function.

Another issue is that `cgroups.GetOwnCgroupPath("cpu")` appears to return a
path like `/sys/fs/cgroup/cpu` instead of the actual cgroup of the cadvisor
process (`/docker/cadvisor`) as expected in the other usages of
`cadvisorContainer`. `cgroups.GetOwnCgroup` returns the correct path.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>