Optimize network metrics collection
Apparently, we collect network stats for *all* containers, and then
discard most (or some) of these statistics:

 - for both crio and containerd, we collect and discard stats for non-infra
   containers containers (i.e. most of containers);

 - for docker, we collect and discard stats for containers which share
   netns with another container (which is rare I guess);

Instead of reading and parsing a bunch of files in /proc/PID/net and
when discarding the just-gathered stats, let's set things up in a way so
we don't collect useless stats in the first place.

This should improve performance, memory usage, and ease the load on
garbage collection.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>