container/common: Replace readUInt64 with optimized version
This patch replaces the `readUint64` function with one that optimizes
for reading uint64s from files only. It prevents unnecessary calls to
`.Stat` because a valid uint64 can be 20 characters at most anyway. It
also avoids a few allocations.

Overall, benchmarking has shown to reduce the total CPU time by ~15%,
which given a total usage of roughly 3% in the entire kubelet, this
means we're reducing the kubelet baseline CPU by about ~0.45%.