Makefile: fix golangci-lint install issues
With commit 300242cc, my intention was to use whatever version of
golangci-lint is available locally, and if not (which is usually the
case in CI), install the version recommended by the Makefile.

The code was changed since, and currently has two issues wrt how
golangci-lint is installed:

1. golangci-lint prints version without v prefix (e.g.
   "... version 1.23.45 ..."), while grep looks for v1.23.45.
   As a result, golangci-lint is *always* reinstalled.

2. If a newer golangci-lint version is installed locally, running
   "make lint" will downgrade it. This is not what most developers
   would expect.

Let's fix both. A version in the Makefile is now minimally
required version, and no attempt to downgrade is made.

Fixes: ea3afbc6 ("Build changes for golang 1.18")
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>