feat: group memory.stats sock metric
This adds the cgroup stat `sock` from the `memory.stats` metric to
cAdvisor.

The motivation is that we've seen numerous examples at DBC Digital of
application developers creating applications that exhaust socket memory,
e.g. by accidentally creating too many TCP connections and not closing
them, or keeping around a few large allocations, or many other such
issues.

Because cAdvisor currently doesn't report socket memory usage, this has
been hard to monitor, and will only be observed when the OOM killer is
reached.

By adding this metric, it will be possible to proactively handle socket
memory exhaustion (which is really kernel memory exhaustion), before it
becomes a potential incident, and to create alerting and enhance
observability of this failure mode.

Signed-off-by: Christina Sørensen <chrs@dbc.dk>