feat: add LoadTaskProcess api in containerd client
This commit aims to add an api to get the task process of the containerd
client.
We have a use case where we are trying to consume the exited containers
and get their exit code however the current implementation of the cliet
dosn't expose this api.
The broader issue and usecase is captured here
https://github.com/google/cadvisor/issues/3722.